Removed TraceDelegate after review
